+++ src/generation/cmark/parse_cmark_ast.rs
@@ -2,6 +2,7 @@ use super::parsing::parse_image as parse_image_from_text;
 use super::parsing::parse_link_reference;
 use super::parsing::parse_link_reference_definitions;
 use crate::generation::common::*;
+use crate::generation::trim_document_whitespace;
 use crate::generation::trim_spaces_and_newlines;
 use crate::generation::trim_start_spaces_and_newlines;
 use pulldown_cmark::*;
@@ -39,10 +40,7 @@ impl<'a> EventIterator<'a> {
       if !self.allow_empty_text_events {
         // skip over any empty text or html events
         while let Some((Event::Text(_), range)) | Some((Event::Html(_), range)) = &self.next {
-          if self
-            .trim_document_whitespace(&self.file_text[range.start..range.end])
-            .is_empty()
-          {
+          if trim_document_whitespace(&self.file_text[range.start..range.end]).is_empty() {
             self.next = self.move_iterator_next();
           } else {
             break;
@@ -53,28 +51,6 @@ impl<'a> EventIterator<'a> {
       Some(event)
     } else {
       None
-    }
-  }
-
-  // This trims only document white space characters as defined in the [Mozilla Developer Network docs](https://developer.mozilla.org/en-US/docs/Web/CSS/Guides/Text/Whitespace#what_is_whitespace).
-  fn trim_document_whitespace<'b>(&self, s: &'b str) -> &'b str {
-    let bytes = s.as_bytes();
-
-    let start = bytes
-      .iter()
-      .position(|&b| !matches!(b, b' ' | b'\t' | b'\r' | b'\n'))
-      .unwrap_or(bytes.len());
-
-    let end = bytes
-      .iter()
-      .rposition(|&b| !matches!(b, b' ' | b'\t' | b'\r' | b'\n'))
-      .map(|i| i + 1)
-      .unwrap_or(0);
-
-    if start <= end {
-      &s[start..end]
-    } else {
-      ""
     }
   }
