+++ src/generation/utils.rs
@@ -99,6 +99,29 @@ pub fn unindent(text: &str) -> Cow<'_, str> {
   }
 }
 
+///  This trims only document white space characters as defined
+/// in the [Mozilla Developer Network docs](https://developer.mozilla.org/en-US/docs/Web/CSS/Guides/Text/Whitespace#what_is_whitespace).
+pub fn trim_document_whitespace<'b>(s: &'b str) -> &'b str {
+  let bytes = s.as_bytes();
+
+  let start = bytes
+    .iter()
+    .position(|&b| !matches!(b, b' ' | b'\t' | b'\r' | b'\n'))
+    .unwrap_or(bytes.len());
+
+  let end = bytes
+    .iter()
+    .rposition(|&b| !matches!(b, b' ' | b'\t' | b'\r' | b'\n'))
+    .map(|i| i + 1)
+    .unwrap_or(0);
+
+  if start <= end {
+    &s[start..end]
+  } else {
+    ""
+  }
+}
+
 pub fn trim_spaces_and_newlines(text: &str) -> &str {
   text.trim_matches(is_space_tab_or_newline)
 }
